Split out OtherCheckbox
