+++ apps/fluent-tester/src/TestComponents/Checkbox/OtherCheckbox.windows.tsx
@@ -1,0 +1,3 @@
+export const OtherCheckbox = () => {
+  return null;
+};
